Introducing the concept of synchronizers...
Terms:
model - a defined structure of some data
working model - a model with methods (like the one in memory which you call in code)
stream model - a model for the serialized data  (like in a file)

Synchronizers should synchronize working models to other models, like stream models.
They are similar to e.g. export/import filters, but stay the whole time until
the working model is closed. They care for everything, like if file/other model is changed,
so the rest of the program does not need to deal with this. Separation of concerns, yeah :)
For synchronizers which synchronize with files in the filesystem there is the 
subclass KAbstractDocumentFileSystemSynchronizer, to care for the general work
with (remote) files.

Other models to synchronize with may also be working models in remote processes/servers.
Like in a collaborative setting. More on this later, not yet included.


svn path=/trunk/playground/utils/okteta/; revision=738470