Added: first approach to a piecetable structure, pretty Leichtwasser-matrosig (works, but has some flaws).
What is a "Piece table":
  stores both original data and a buffer with all new input,
  current data is defined by a list of pieces pointing to offsets in one
  of the buffers and some length. Changes are done by appending to the
  input buffer if any and manipulating the list of pieces.
  In use by Abiword, Qt 4 QTextDocument and others.

And it is an Alpha version of the code, to get things at least working,
don't judge too quickly, please. Help is welcome :)

It was tried to have it to be ignorant of the supported data structure,
only dependencies so far are primitive utility classes KSection and Co,
no used knowledge of KAbstractByteArrayModel or else, cmp. namespace KHE
vs. namespace KHECore.
So other interested parties might try to make use of this code, too. They
are welcomed.

But the undo/redo stuff, aka revertable changes, seems to force one to finally
use structures from the basic data model, e.g. otherwise a list of piecetable
changes would have to been transformed into a list of bytearray changes.
And no, currently I do not want to rely the KAbstractByteArrayModel on
the piecetable library, neither. Perhaps a common usage can be done by turning the
piecetable lib into template code...


svn path=/trunk/playground/utils/okteta/; revision=775829