Moving Okteta from kdereview to kdeutils.
The two weeks in kdereview passed without any objection,
also the module coordinator is okay with it, (who's me, after all),
so Okteta is in. That is, kdeutils.

The code in kdeutils/okteta/{core,gui,parts} obsoletes the one
in kdeutils/khexedit/{core,gui,parts}, as it was branched
from the latter one before, so that one is removed now.
And such the khexedit directory is completely empty and is removed, too,
because the code for the program KHexEdit was removed as unmaintained before
the 4.0 release.

[memorial minute]You will stay in many people's memory, KHexEdit.[/memorial minute]

But hopefully Okteta will be better soon, so let's go on now :)

Please, could someone from the translation team move the translations, too?
Thanks a lot.

CCMAIL:release-team@kde.org
CCMAIL:kde-i18n-doc@kde.org


svn path=/trunk/KDE/kdeutils/okteta/; revision=797377