added: preliminary job concept to all io-related processes, to fix blocked ui updates
(or: even more classes, to ensure the leadership at the classes/locs index)

With 4.2 the support for asynchronic processes will be introduced. This commit just
brings some first basics hacked to run all jobs synchronously and have the ui updated
if needed, as a fix for 4.1. Really, hacked at a lot of places.
So do not wonder that jobs, while having already worker threads, wait in an
own loop for the thread to finish. It just works for now (tm).

Looks pretty invasive, but the jobs and threads got almost all of their code/algorithms
from the synchronizer classes, so in fact this is not such a big change, more kind
of a refactoring. And guarded by unit tests, which do not fail. :)


svn path=/trunk/KDE/kdeutils/okteta/; revision=825690