Backport 1218154:
Move the call to setFixedFontByGlobalSettings() to right after the QTreeView is initialized. This ensures the font is set before the model and therefore avoids calculating the width of everything in the model.

In my tests this improved the startup time from just over 2 secs to round about 1 sec (Callgrind says from 8 billion instr down to 4)

svn path=/branches/KDE/4.6/kdesdk/okteta/; revision=1218155