Further improve startup speed of Okteta.
This commit changes the way the width of the columns in the tree views are calculated. Now they are interactively resizable and not ResizeToContents which is very expensive to calculate.

On my system this reduces the time needed to start Okteta from ~1 seconds to under 500 milliseconds, roughly a fifth of what it was before this patch and the previous one.

svn path=/trunk/KDE/kdesdk/okteta/; revision=1218164