Reduce memory usage of the structures tool.
Previously the current value for each element was stored in a union (64 bits) so that the field could be in the superclass.
Now every subclass for each primitive data type has its own member variable of the fitting size. I.e. if there is a 150 char string withing the structure this already saves ~1K of memory (1050 bytes exactly)

svn path=/trunk/KDE/kdesdk/okteta/; revision=1218518