Add a new field to DataInformation to save the byte order
Following values are possible: FromSettings, Inherit, LittleEndian, BigEndian. This saved value is then used to read the data instead of always the value from the setting.
This will allow updating some structures - like png which is always big endian - so that they are correct regardless of the current endianess settings and also allow structures containing values of mixed byte order.

This commit also fixes an issue where a value could be written to the wrong offset in the file (only with union members)


svn path=/trunk/KDE/kdesdk/okteta/; revision=1218701