Allow setting the byte order of each element in the XML structure definitions.
An extra attribute is added to all elements named "byteOrder".
The following values are accepted: "inherit" (use value from parent, or if there is no parent the default setting), "fromSettings" (use the value from the settings), "littleEndian", "bigEndian". The value is compared case-insensitively, so even "BiGENDian" will be accepted.

If no value is set it defaults to "inherit".

CCBUG: 263496

svn path=/trunk/KDE/kdesdk/okteta/; revision=1218708