Also allow setting the byte order for structures defined in JavaScript. All that has to be done is set the property "byteOrder" of the chosen element to "inherit", "fromSettings", "littleEndian" or "bigEndian" (case insensitive, any invalid values will be "inherit").
I did not add it to the constructor since remembering the order of parameters is not very nice. The following has to be done to set the byte order:

Suppose this is our init() function:

funtion init() {
  var obj = struct({aFloat : float(),
      charArray : array(int32(),4),
      unsignedInt : uint32(),
      bitField : bitfield("unsigned",10)});
  obj.unsignedInt.byteOrder = "littleEndian";
  return obj;
}

This makes unsignedInt use little endian regardless of the byte order from the settings, all others use the byte order from the settings.

CCBUG: 263496

svn path=/trunk/KDE/kdesdk/okteta/; revision=1218713