Improve parsing of enum values from JavaScript Now not only uint32 literals are possible, but larger values too.
However you have to ensure that the value is written as a string instead
of a number literal for values greater than 9007199254740992 as JS
stores numbers a a double internally and after that value it is not
possible to represent every integer value.

Also it is no longer possible to have doubles/floats as enum constants
If there is one, it is silently converted to an integer.

svn path=/trunk/KDE/kdesdk/okteta/; revision=1275910