structures: refactor AdditionalData
It is now a just a QHash. This makes it easy to store custom data in it
Also makes using it much easier since it is stored by value instead of
as a pointer. Memory usage when not needed is still the same since
sizeof(empty QHash) == sizeof(void*). Memory usage when needed may be a
bit higher, but since it will only be needed rarely this is not an issue