No longer compile with -fno-rtti
Otherwise we get a crash in KConfigDialogManager::setupWidget which uses
dynamic_cast<>. Generally this is quite risky since the other libraries
might be using RTTI features. This used to work since the dynamic_cast
was only added to KConfigDialogManager recently. Keep RTTI on now so
that something like this doesn't happen anymore (premature optimization)