Fix assertion when changing selected structure
If the previous structure was locked to an offset and the current one
isn't it would attempt to unlock the current structure since the state
of the button had changed. This would assert since the current one isn't
locked. Fixed by using a QSignalBlocker whenever the selection changes
so that the toggled() signal is only emitted whenever the user clicks on
the button.

CCMAIL: ps_ml@gmx.de