Remove the .sh suffix from struct2osd
This should fix the following lintian error:
When scripts are installed into a directory in the system PATH, the
script name should not include an extension such as .sh or .pl that
denotes the scripting language currently used to implement it. The
implementation language may change; if it does, leaving the name the
same would be confusing and changing it would be disruptive.

BUG: 345092