Primitive array element type consistency.
In a primitive array, defer to the primitive child type for byteOrder, typeName, and creating a value string for the UI. This allows user defined types which are based on primitive types but override these properties to be displayed as the user defined types otherwise would when used stand-alone.

Testing Done:
Built, built the 'test' target (passed tests), installed, ran with a main.js like the following attached file on a valid .ttf file. Observed that with this change the 'cvt ' array entries have the expected endianness, type name, and value string.

REVIEW: 125523