XFAIL two tests on macOS and Windows
I tried to add a way of adding additional data directories in
upstream Qt (https://codereview.qt-project.org/c/qt/qtbase/+/239952)
a while ago but that change was vetoed, so XFAIL the test for now.
The same logic is probably required on Windows, but I don't have
a system to test this.