Use "" for includes of headers of the own library
POSIX at least for C compilers defines that "" includes
"shall be searched for first in the directory of the file with the
#include line", overruling the "The named source file is
searched for in an implementation-defined manner." of the C standard.

Same behaviour seems also common for preprocessing of C++ compilers,
so rely on this consistently.

Advantages:
* shorter include statements due to no need of namespace dir prefix
* less file system stats needed to resolve a header (µ-opt)

GIT_SILENT