Fix initial too large window size
Adding all the dock widgets next to each other on the right dock area at
initialization results in an initial minimal size of the main window that
is quite large vertically (e.g. > 2000 px). As KXmlGuiWindow::setupGUI()
first tries to restore the size before trying to restore the positions of
the dock widgets, any stored size will be minimally bounded by that large
initial minimal size. As the stored size usually was smaller, due to users
typically hiding or stacking dock widgets for some usable window size
(similar also the default state set in OktetaMainWindow constructor), this
effectively discards the stored size, also results in a vertical start
size which does not fit on most typical consumer screens.
On KWin/X11 this effect was shadowed a bit by (assumingly) the window
manager trying to fit the window on start into the screen, at which stage
the dock widgets positions had already been restored and thus the minimal
size being typically small enough again to fit into the screen.

Not adding the widgets to the dock area initially does not prevent them
from being correctly restored later if the stored config asks for it.
Because setting the main window as parent widget to the dock widget makes
it known to the dock widget system.

BUG: 462703
FIXED-IN: 0.26.17